+++ Source/AccelByteUe4Sdk/Private/Core/AccelByteMessagingSystem.cpp
@@ -22,6 +22,7 @@ FAccelByteMessagingSystem::FAccelByteMessagingSystem()
 
 FAccelByteMessagingSystem::~FAccelByteMessagingSystem()
 {
+	UnsubscribeAll();
 	FTickerAlias::GetCoreTicker().RemoveTicker(PollHandle);
 }